Move clean shutdown file to storage package
