+++ checkstyle/import-control-storage.xml
@@ -75,6 +75,7 @@
     </subpackage>
 
     <subpackage name="storage.internals">
+        <allow pkg="com.fasterxml.jackson" />
         <allow pkg="com.yammer.metrics.core" />
         <allow pkg="org.apache.kafka.server"/>
         <allow pkg="org.apache.kafka.storage.internals"/>
